Refactoring db utils
