+++ server/controllers/userController.js
@@ -1,4 +1,4 @@
-const db = require('../db/db');
+const db = require('../db');
 const UserModel = require('../models/userModel');
 
 const userController = {};
