+++ server/utils/scripts/createTables.js
@@ -1,0 +1,6 @@
+const Query = require('../PgBaseUtils');
+const data = require('./createData');
+
+module.exports = (function create() {
+    Query.createTable(data.tablesData);
+}());
